+++ Environment_Engine/Query/PointsInSpaces.cs
@@ -1,4 +1,4 @@
-/*
+﻿/*
  * This file is part of the Buildings and Habitats object Model (BHoM)
  * Copyright (c) 2015 - 2020, the respective contributors. All rights reserved.
  *
@@ -39,21 +39,16 @@ namespace BH.Engine.Environment
         /**** Public Methods                            ****/
         /***************************************************/
 
-        [Description("Returns a BHoM Geometry Point that is in the provided space")]
-        [Input("panelsAsSpace", "The collection of Environment Panels that represent a single space to get the point in")]
-        [Output("spacePoint", "A point in the space")]
-        public static Point PointInSpace(this List<Panel> panelsAsSpace)
+        [Description("Returns a collection of BHoM Geometry Points that are in each space")]
+        [Input("panelsAsSpaces", "The nested collection of Environment Panels that represent the spaces to points in")]
+        [Output("spacePoints", "A collection of points in each space")]
+        public static List<Point> PointsInSpaces(this List<List<Panel>> panelsAsSpaces)
         {
-            Polyline floor = panelsAsSpace.FloorGeometry();
-            Point spacePnt = floor.PointInRegion();
+            List<Point> spacePnts = new List<Point>();
+            foreach (List<Panel> space in panelsAsSpaces)
+                spacePnts.Add(space.PointInSpace());
 
-            double height = 0.0;
-            foreach (Panel p in panelsAsSpace)
-                height = Math.Max(height, p.Height());
-
-            spacePnt.Z += height / 2;
-
-            return spacePnt;
+            return spacePnts;
         }
     }
 }
